display the raw ingredient ref during load
This is super fast when the server's hot, but it's now possible to do
a client-side load of a detail page, without having the ingredients
available, and have to wait for a container to spin up to get them.

By starting with the raw version, you can at least look at the recipe
during spin up, even if it's not as pretty.