avoid needless re-recog on draft ingredients
Using an index key means when an ingredient is inserted or deleted on
the draft recipe form all later ingredients are completely redrawn, and
therefore re-recognize their (unchanged) data.

By using a ClientId for the key, that is avoided. The id isn't used for
server sync as on the planner, as recipe ingredients don't have "real"
identity.