"address" much CRA/Node/NPM stupidity
Since early last year, when CRA does a hot reload, its error overlay
drops an iframe over the entire document, forcing a manual reload. The
kludge of forcing an old version of the overlay seems to make it less
stupid (https://github.com/facebook/create-react-app/issues/11880).

Switch to using CRA's eslint stuff, which was sorta equivalent-ish,
after adding our existing config to package.json. Still required a
bunch of no-op twiddles to get clean again, as well as disabling
linting for LoadObject and friends. They're vendored, Flow (not JS),
and on their way out.

Next, upgrade react-showdown, as the version we were using doesn't
work with Webpack 5, which react-scripts switched to (from 4).

Finally, MaterialUI got an unrequested update, which deprecated the
theme builder function. That's spiffed up. I didn't go digging as to
where/why it changed, or if there are other things. MUI seems to be
pretty good with semver, which is a refreshing change.

Some of the above glergage would be simplified with Yarn's PnP, as it
offers more control over the transitive stuff, but part of it is due to
CRA itself. And Node itself is at the root of the blame tree.