Sending to plan no longer guarantees recipe loaded
When the library was searched via REST, it was impossible to trigger a
send-to-plan for a recipe not already in the store. But with search
using GraphQL, buttons on the result cards may not correspond to stored
recipes. So gracefully fallback to "recipe" in the snackbar alert, if
the store doesn't have the recipe being sent.