with react 18, vite dev mode doesn't like callback styles
This is a weird one. With react 17, vite's dev mode was happy with using
callbacks w/in makeStyles. With react 18, however, the initial render is
missing any callback-based styles. If HMR fires for the component, then
subsequent renders will be correct, but the first one is broken. The
built version doesn't have the issue, only dev mode. PageBody is the
only instance of a style callback, so use an additional class to avoid
the callback.